+++ ending-credits-api/src/main/java/com/hanaro/endingcredits/endingcreditsapi/utils/adapter/dto/clova/OCR/id/ClovaDLDto.java
@@ -1,4 +1,4 @@
-package com.hanaro.endingcredits.endingcreditsapi.utils.adapter.dto.clova.OCR;
+package com.hanaro.endingcredits.endingcreditsapi.utils.adapter.dto.clova.OCR.id;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,10 +11,15 @@ import java.util.List;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ClovaICDto {
+public class ClovaDLDto {
+    private List<Object> type;
+    private List<Object> num;
     private List<ClovaIdCardDetailDto> name;
     private List<ClovaIdCardDetailDto> personalNum;
     private List<ClovaIdCardDetailDto> address;
+    private List<Object> renewStartDate;
+    private List<Object> renewEndDate;
+    private List<Object> code;
     private List<Object> issueDate;
     private List<Object> authority;
 }
