+++ ending-credits-api/src/main/java/com/hanaro/endingcredits/endingcreditsapi/utils/adapter/dto/clova/OCR/id/ClovaICImageDto.java
@@ -1,4 +1,4 @@
-package com.hanaro.endingcredits.endingcreditsapi.utils.adapter.dto.clova.OCR;
+package com.hanaro.endingcredits.endingcreditsapi.utils.adapter.dto.clova.OCR.id;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,11 +9,11 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ClovaDLImageDto {
+public class ClovaICImageDto {
     private String uid;
     private String name;
     private String inferResult;
     private String message;
     private Object validationResult;
-    private ClovaDLIdCardDto idCard;
+    private ClovaICIdCardDto idCard;
 }
